AG-16452 - Break out benchmark.js.
