+++ plugins/ag-charts-generate-example-files/src/executors/generate/generator/utils/getBenchmarkSnippet.ts
@@ -3,8 +3,8 @@ export const BENCHMARK_END = '/** BENCHMARK END **/';
 
 export const getBenchmarkSnippet = () =>
     `${BENCHMARK_START}
-// Benchmark loader - imports and initializes the benchmark runner
-import('./benchmark.js').then(({ initBenchmark }) => {
+// Benchmark loader - imports and initializes the benchmark harness
+import('./benchmarkHarness.js').then(({ initBenchmark }) => {
     if (typeof getBenchmarkConfig === 'function') {
         initBenchmark(getBenchmarkConfig());
     }
